commentaires et polissage

+++ combination.h
@@ -1,3 +1,3 @@
 char** combination_arrays(char **key_lenght, int n, int *cursor, char **c_key, int int_key_lenght);
-char** get_key_list(char **key_lenght, char **c_key, unsigned int *key_list_n);
+char** get_key_list(char **key_lenght, char **c_key, unsigned long *key_list_n);
